refactor: Simplify release CI workflow
