--- .github/actions/release-notes/action.yml
@@ -1,23 +1,0 @@
-name: 'Release Notes'
-description: 'Return release notes based on Git Commits'
-inputs:
-  from:
-    description: 'Commit from which start log'
-    required: true
-  to:
-    description: 'Commit to which end log'
-    required: true
-  include-commit-body:
-    description: 'Should the commit body be in notes'
-    required: false
-    default: 'false'
-  include-abbreviated-commit:
-    description: 'Should the commit sha be in notes'
-    required: false
-    default: 'true'
-outputs:
-  release-note: # id of output
-    description: 'Release notes'
-runs:
-  using: 'node12'
-  main: 'main.js'
