+++ .github/workflows/release.yml
@@ -1,5 +1,16 @@
 name: Release
-on: [ workflow_call, workflow_dispatch ]
+on:
+  workflow_dispatch:
+    inputs:
+      version:
+        required: true
+        type: string
+        description: Semantic version of new release. Without `v` prefix.
+  workflow_call:
+    inputs:
+      version:
+        required: true
+        type: string
 
 concurrency:
   group: release-${{ github.ref }}
@@ -12,57 +23,7 @@ defaults:
 
 
 jobs:
-
-  draft:
-    runs-on: ubuntu-latest
-    outputs:
-      release-note: ${{ steps.release-note.outputs.release-note }}
-      version: ${{ steps.version.outputs.build-version }}
-
-    steps:
-      - uses: actions/checkout@v3
-        with:
-          fetch-depth: 0
-
-      - uses: actions/setup-node@v3
-        with:
-          node-version: 14
-
-      - name: Get last git tag
-        id: tag
-        run: echo "::set-output name=last-tag::$(git describe --tags --abbrev=0 || git rev-list --max-parents=0 ${{github.ref}})"
-
-      - name: Generate release notes
-        uses: ./.github/actions/release-notes
-        id: release-note
-        with:
-          from: ${{ steps.tag.outputs.last-tag }}
-          to: ${{ github.ref }}
-          include-commit-body: true
-          include-abbreviated-commit: true
-
-      - name: Get version from current date
-        id: version
-        run: echo "::set-output name=build-version::$(node -e "try{console.log(require('./.electron-builder.config.js').extraMetadata.version)}catch(e){console.error(e);process.exit(1)}")"
-
-      - name: Delete outdated drafts
-        uses: hugo19941994/delete-draft-releases@v1.0.0
-        env:
-          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-
-      - name: Create Release Draft
-        uses: softprops/action-gh-release@v1
-        env:
-          GITHUB_TOKEN: ${{ secrets.github_token }}
-        with:
-          prerelease: true
-          draft: true
-          tag_name: v${{ steps.version.outputs.build-version }}
-          name: v${{ steps.version.outputs.build-version }}
-          body: ${{ steps.release-note.outputs.release-note }}
-
   upload_artifacts:
-    needs: [ draft ]
 
     strategy:
       matrix:
@@ -84,20 +45,13 @@ jobs:
         env:
           PLAYWRIGHT_SKIP_BROWSER_DOWNLOAD: 1
 
-      # The easiest way to transfer release notes to a compiled application is create `release-notes.md` in the build resources.
-      # See https://github.com/electron-userland/electron-builder/issues/1511#issuecomment-310160119
-      - name: Prepare release notes
-        env:
-          RELEASE_NOTE: ${{ needs.draft.outputs.release-note }}
-        run: echo "$RELEASE_NOTE" >> ./buildResources/release-notes.md
-
       # Compile app and upload artifacts
       - name: Compile & release Electron app
         uses: samuelmeuli/action-electron-builder@v1
         env:
-          VITE_APP_VERSION: ${{ needs.draft.outputs.version }}
+          VITE_APP_VERSION: ${{ inputs.version }}
         with:
-          build_script_name: build
+          build_script_name: build # npm run build
           args: --config .electron-builder.config.js
 
           # GitHub token, automatically provided to the action
