+++ packages/preload/vite.config.js
@@ -2,6 +2,10 @@ import {chrome} from '../../.electron-vendors.cache.json';
 import {preload} from 'unplugin-auto-expose';
 import {join} from 'node:path';
 
+if (process.env.VITE_APP_VERSION === undefined) {
+  throw new Error('VITE_APP_VERSION Environment variable is required');
+}
+
 const PACKAGE_ROOT = __dirname;
 const PROJECT_ROOT = join(PACKAGE_ROOT, '../..');
