+++ packages/renderer/vite.config.js
@@ -5,6 +5,10 @@ import vue from '@vitejs/plugin-vue';
 import {renderer} from 'unplugin-auto-expose';
 import {join} from 'node:path';
 
+if (process.env.VITE_APP_VERSION === undefined) {
+  throw new Error('VITE_APP_VERSION Environment variable is required');
+}
+
 const PACKAGE_ROOT = __dirname;
 const PROJECT_ROOT = join(PACKAGE_ROOT, '../..');
